reorganize Rust modules
