+++ opening-hours-py/src/types/location.rs
@@ -1,6 +1,6 @@
 use chrono::{NaiveDate, NaiveDateTime};
 
-use opening_hours::{Localize, NoLocation, TzLocation};
+use opening_hours::localization::{Localize, NoLocation, TzLocation};
 use opening_hours_syntax::rules::time::TimeEvent;
 
 use super::datetime::DateTimeMaybeAware;
