+++ opening-hours/benches/benchmarks.rs
@@ -1,4 +1,4 @@
-use opening_hours::country::Country;
+use opening_hours::localization::{Coordinates, Country};
 use opening_hours::{Context, OpeningHours};
 
 use chrono::NaiveDateTime;
@@ -8,7 +8,7 @@ const SCH_24_7: &str = "24/7";
 const SCH_ADDITION: &str = "10:00-12:00 open, 14:00-16:00 unknown, 16:00-23:00 closed";
 const SCH_HOLIDAY: &str = "PH";
 const SCH_JAN_DEC: &str = "Jan-Dec";
-const PARIS_COORDS: [f64; 2] = [48.8535, 2.34839];
+const PARIS_COORDS: Coordinates = Coordinates::new(48.8535, 2.34839).unwrap();
 
 fn bench_parse(c: &mut Criterion) {
     let mut group = c.benchmark_group("parse");
@@ -26,7 +26,7 @@ fn bench_context(c: &mut Criterion) {
     let mut group = c.benchmark_group("context");
 
     group.bench_function("infer_from_coords", |b| {
-        b.iter(|| Context::from_coords(black_box(PARIS_COORDS[0]), black_box(PARIS_COORDS[1])))
+        b.iter(|| Context::from_coords(black_box(PARIS_COORDS)))
     });
 }
