+++ opening-hours/src/context.rs
@@ -1,11 +1,9 @@
 use std::fmt::Debug;
-use std::ops::Add;
 use std::sync::Arc;
 
-use chrono::{Datelike, Duration, NaiveDate, NaiveDateTime, NaiveTime, TimeDelta, TimeZone};
 use compact_calendar::CompactCalendar;
-use opening_hours_syntax::rules::time::TimeEvent;
-use sunrise_next::{DawnType, SolarDay, SolarEvent};
+
+use crate::localization::{Localize, NoLocation};
 
 // --
 // -- Holidays
@@ -35,224 +33,6 @@ impl ContextHolidays {
     }
 }
 
-// --
-// -- Coordinates
-// --
-
-/// A valid pair of geographic coordinates.
-///
-/// See https://en.wikipedia.org/wiki/Geographic_coordinate_system
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd)]
-pub struct Coordinates {
-    lat: f64,
-    lon: f64,
-}
-
-impl Coordinates {
-    /// Validate a pair of latitude / longitude.
-    ///
-    /// Return `None` if values are out of range (`abs(lat) > 90` or
-    /// `abs(lon) > 180`).
-    pub const fn new(lat: f64, lon: f64) -> Option<Self> {
-        if lat < -90.0 || lat > 90.0 || lon < -180.0 || lon > 180.0 {
-            return None;
-        }
-
-        Some(Self { lat, lon })
-    }
-
-    /// Get the time for a sun event at a given date.
-    pub fn event_time(&self, date: NaiveDate, event: TimeEvent) -> chrono::DateTime<chrono::Utc> {
-        let solar_event = match event {
-            TimeEvent::Dawn => SolarEvent::Dawn(DawnType::Civil),
-            TimeEvent::Sunrise => SolarEvent::Sunrise,
-            TimeEvent::Sunset => SolarEvent::Sunset,
-            TimeEvent::Dusk => SolarEvent::Dusk(DawnType::Civil),
-        };
-
-        let solar_day = SolarDay::new(self.lat, self.lon, date.year(), date.month(), date.day());
-        let timestamp = solar_day.event_time(solar_event);
-        chrono::DateTime::from_timestamp(timestamp, 0).expect("invalid timestamp")
-    }
-
-    /// Get latitude component.
-    pub fn lat(&self) -> f64 {
-        self.lat
-    }
-
-    /// Get longitude component.
-    pub fn lon(&self) -> f64 {
-        self.lon
-    }
-}
-
-impl std::fmt::Display for Coordinates {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "({}, {})", self.lat, self.lon)
-    }
-}
-
-// --
-// -- Localization
-// --
-
-/// Specifies how dates should be localized while evaluating opening hours. No
-/// localisation is available by default but this can be used to specify a
-/// timezone and coordinates (which affect sun events).
-pub trait Localize: Clone + Send + Sync {
-    /// The type for localized date & time.
-    type DateTime: Clone + Add<Duration, Output = Self::DateTime>;
-
-    /// Get naive local time.
-    fn naive(&self, dt: Self::DateTime) -> NaiveDateTime;
-
-    /// Localize a naive datetime.
-    fn datetime(&self, naive: NaiveDateTime) -> Self::DateTime;
-
-    /// Get the localized time for a sun event at a given date.
-    fn event_time(&self, _date: NaiveDate, event: TimeEvent) -> NaiveTime {
-        match event {
-            TimeEvent::Dawn => NaiveTime::from_hms_opt(6, 0, 0).unwrap(),
-            TimeEvent::Sunrise => NaiveTime::from_hms_opt(7, 0, 0).unwrap(),
-            TimeEvent::Sunset => NaiveTime::from_hms_opt(19, 0, 0).unwrap(),
-            TimeEvent::Dusk => NaiveTime::from_hms_opt(20, 0, 0).unwrap(),
-        }
-    }
-}
-
-// No location info.
-#[derive(Clone, Debug, Default, Hash, PartialEq, Eq)]
-pub struct NoLocation;
-
-impl Localize for NoLocation {
-    type DateTime = NaiveDateTime;
-
-    fn naive(&self, dt: Self::DateTime) -> NaiveDateTime {
-        dt
-    }
-
-    fn datetime(&self, naive: NaiveDateTime) -> Self::DateTime {
-        naive
-    }
-}
-
-/// Time zone is specified and coordinates can optionally be specified for
-/// accurate sun events.
-#[derive(Clone, Debug, PartialEq)]
-pub struct TzLocation<Tz>
-where
-    Tz: TimeZone + Send + Sync,
-{
-    tz: Tz,
-    coords: Option<Coordinates>,
-}
-
-impl<Tz> TzLocation<Tz>
-where
-    Tz: TimeZone + Send + Sync,
-{
-    /// Create a new location context which only contains timezone information.
-    pub fn new(tz: Tz) -> Self {
-        Self { tz, coords: None }
-    }
-
-    /// Extract the timezone for this location.
-    pub fn get_timezone(&self) -> &Tz {
-        &self.tz
-    }
-
-    /// Attach coordinates to the location context.
-    ///
-    /// If coordinates where already specified, they will be replaced with the
-    /// new ones.
-    pub fn with_coords(self, coords: Coordinates) -> Self {
-        Self { tz: self.tz, coords: Some(coords) }
-    }
-}
-
-#[cfg(feature = "auto-timezone")]
-impl TzLocation<chrono_tz::Tz> {
-    /// Create a new location context from a set of coordinates and with timezone
-    /// information inferred from this localization.
-    ///
-    /// Returns `None` if latitude or longitude is invalid.
-    ///
-    /// ```
-    /// use chrono_tz::Europe;
-    /// use opening_hours::{Coordinates, TzLocation};
-    ///
-    /// let coords = Coordinates::new(48.8535, 2.34839).unwrap();
-    ///
-    /// assert_eq!(
-    ///     TzLocation::from_coords(coords),
-    ///     TzLocation::new(Europe::Paris).with_coords(coords),
-    /// );
-    /// ```
-    pub fn from_coords(coords: Coordinates) -> Self {
-        use std::collections::HashMap;
-        use std::sync::LazyLock;
-
-        static TZ_NAME_FINDER: LazyLock<tzf_rs::DefaultFinder> =
-            LazyLock::new(tzf_rs::DefaultFinder::new);
-
-        static TZ_BY_NAME: LazyLock<HashMap<&str, chrono_tz::Tz>> = LazyLock::new(|| {
-            chrono_tz::TZ_VARIANTS
-                .iter()
-                .copied()
-                .map(|tz| (tz.name(), tz))
-                .collect()
-        });
-
-        let tz_name = TZ_NAME_FINDER.get_tz_name(coords.lon, coords.lat);
-
-        #[allow(clippy::unnecessary_lazy_evaluations)]
-        let tz = TZ_BY_NAME.get(tz_name).copied().unwrap_or_else(|| {
-            #[cfg(feature = "log")]
-            log::warn!("Could not find time zone `{tz_name}` at {coords}");
-            chrono_tz::UTC
-        });
-
-        Self::new(tz).with_coords(coords)
-    }
-}
-
-impl<Tz> Localize for TzLocation<Tz>
-where
-    Tz: TimeZone + Send + Sync,
-    Tz::Offset: Send + Sync,
-{
-    type DateTime = chrono::DateTime<Tz>;
-
-    fn naive(&self, dt: Self::DateTime) -> NaiveDateTime {
-        dt.with_timezone(&self.tz).naive_local()
-    }
-
-    fn datetime(&self, mut naive: NaiveDateTime) -> Self::DateTime {
-        loop {
-            if let Some(dt) = self.tz.from_local_datetime(&naive).latest() {
-                return dt;
-            }
-
-            naive = naive
-                .checked_add_signed(TimeDelta::minutes(1))
-                .expect("no valid datetime for time zone");
-        }
-    }
-
-    fn event_time(&self, date: NaiveDate, event: TimeEvent) -> NaiveTime {
-        let Some(coords) = self.coords else {
-            return NoLocation.event_time(date, event);
-        };
-
-        let dt = coords.event_time(date, event).with_timezone(&self.tz);
-        self.naive(dt).time()
-    }
-}
-
-// --
-// -- Context
-// --
-
 /// All the context attached to a parsed OpeningHours expression and that can
 /// alter its evaluation semantics.
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]
@@ -274,13 +54,13 @@ impl<L> Context<L> {
 }
 
 #[cfg(feature = "auto-timezone")]
-impl Context<TzLocation<chrono_tz::Tz>> {
+impl Context<crate::localization::TzLocation<chrono_tz::Tz>> {
     /// Create a context with given coordinates and try to infer a timezone and
     /// a local holiday calendar.
     ///
     /// ```
-    /// use opening_hours::{Context, Coordinates, TzLocation};
-    /// use opening_hours::country::Country;
+    /// use opening_hours::Context;
+    /// use opening_hours::localization::{Coordinates, Country, TzLocation};
     ///
     /// let coords = Coordinates::new(48.8535, 2.34839).unwrap();
     ///
@@ -292,14 +72,14 @@ impl Context<TzLocation<chrono_tz::Tz>> {
     /// );
     /// ```
     #[cfg(feature = "auto-country")]
-    pub fn from_coords(coords: Coordinates) -> Self {
-        use crate::country::Country;
+    pub fn from_coords(coords: crate::localization::Coordinates) -> Self {
+        use crate::localization::Country;
 
         let holidays = Country::try_from_coords(coords)
             .map(Country::holidays)
             .unwrap_or_default();
 
-        let locale = TzLocation::from_coords(coords);
+        let locale = crate::localization::TzLocation::from_coords(coords);
         Self { holidays, locale }
     }
 }
