+++ opening-hours/src/lib.rs
@@ -4,14 +4,13 @@
 #![allow(unexpected_cfgs)]
 #![cfg_attr(all(doc, CHANNEL_NIGHTLY), feature(doc_auto_cfg))]
 
-pub mod date_filter;
 #[macro_use]
 pub mod schedule;
-pub mod context;
-pub mod country;
-pub mod opening_hours;
-pub mod time_filter;
+pub mod localization;
 
+mod context;
+mod filter;
+mod opening_hours;
 mod utils;
 
 #[cfg(test)]
@@ -19,7 +18,7 @@ mod tests;
 
 // Public re-exports
 // TODO: make opening_hours.rs lighter and less spaghetty
-pub use crate::context::{Context, ContextHolidays, Coordinates, Localize, NoLocation, TzLocation};
-pub use crate::opening_hours::OpeningHours;
+pub use crate::context::{Context, ContextHolidays};
+pub use crate::opening_hours::{OpeningHours, DATE_LIMIT};
 pub use crate::utils::range::DateTimeRange;
 pub use opening_hours_syntax::rules::RuleKind;
